UPD 21.02.23 -> опции и команды теперь разделены по разным файлам и папкам, для них созданы 2-е глобальных переменные COMMANDS, OPTIONS. Команды регистрируем в файле /cmds/index.mjs. Опции регистрируем в массиве объектов /cmds/options/options.mjs. Небольшие правки в спавнере компонентов, теперь будет проверяться пришедшее значение из строки, если undefined то будет выводиться сообщение об ошибке если всё нормально, то будет создаваться компонент с переданным именем.
