+++ README.md
@@ -13,7 +13,8 @@
 
 ***
 ## Папка проекта
-* по умолчанию папка без компонентов выглядит так:
+* Предпологается что папка с проектом будет как на примере ниже:
+* *если ваш проект выглядит иначе, вы всегда можете поменять пути в скриптах: **componentSpawner.mjs** и в **injectStyles.mjs***
 
 ```
 src
@@ -52,16 +53,16 @@ src/templates
   * ```npm i```
 * После установки зависимостей можно использовать.
 * Для просмотра команд можно ввести просто:
-* ```node app.mjs```
+* ```creat-component-app```
 * Чтобы создать компонент, необходимо вызвать следующие команды:
-  * ```node app.mjs create -n component_name```
+  * ```creat-component-app create -n component_name```
   * или
-  * ```node app.mjs create --name component_name```
+  * ```creat-component-app create --name component_name```
 * Функция подставит название которое вы передали в командной строке и создаст на его основе необходимый набор файлов и папок. 
-* Чтобы добавить ```@import '';``` до компонента в главный файл стилей необходимо ввести имя компонента, если скрипт его найдёт то добавит импорт стилей компонента в главный файл стилей, в противом случае выведет информацию об ошибке в консоль. Воспользуйтесь следующими командами:
-  * ```node app.mjs inject --inject component_name```
+* Чтобы добавить ```@import 'componentName.scss';``` до компонента в главный файл стилей необходимо ввести имя компонента, если скрипт его найдёт то добавит импорт стилей компонента в главный файл стилей, в противом случае выведет информацию об ошибке в консоль. Воспользуйтесь следующими командами:
+  * ```creat-component-app import -s component_name```
   * или
-  * ```node app.mjs inject --style component_name```
+  * ```creat-component-app import --style component_name```
 
 <br>
 
@@ -81,5 +82,12 @@ src/templates
 * Маски выглядят следующим образом:
   * Директории: ```./${name}/```
   * Файлы: ```${name}.njk```
+
+***
+
+***
+
+## Удобства
+- Для удобства, опции и команды отделены.
 
 ***
