--- app.mjs
@@ -1,5 +1,0 @@
-'use strict';
-
-import yargsCliApp from "./.cli/creat-component-app.mjs";
-
-yargsCliApp();
