+++ cmds/utils/injectStyles.mjs
@@ -10,19 +10,21 @@ import chalk from 'chalk';
 const __dirname = path.resolve();
 
 
-const injectStyle = async (inject) => {
+const injectStyle = async (style) => {
   try {
+    const value = style; // переопределяем переменную, для красоты кода.
+
     const componentsPath = './src/templates/components';
     const assetsPath = './src/assets/styles';
     const importPath = '../../templates/components';
 
-    const findDir = `${componentsPath}/${inject}`;
+    const findDir = `${componentsPath}/${value}`;
 
     // здесь жёстко фиксируем путь до файла в который будем добавлять импорт стилей
     const injectTo = `${assetsPath}/main.scss`;
 
     // здесь подставляем максимально достоверный путь до файла стилей компонента.
-    const styleImport = `\n@import '${importPath}/${inject}/${inject}.scss';`;
+    const styleImport = `\n@import '${importPath}/${value}/${value}.scss';`;
 
     // ищем директорию с компонентом, имя получаем из командной.
     await fs.readdir(path.resolve(__dirname, findDir), 'utf8', (err, files) => {
@@ -38,14 +40,14 @@ const injectStyle = async (inject) => {
         }
 
         // проверяем является ли файл стилей, файлом если да, импортируем его в главный файл стилей.
-        fs.stat(`${findDir}/${inject}.scss`, (errStatus, status) => {
+        fs.stat(`${findDir}/${value}.scss`, (errStatus, status) => {
           if (errStatus) {
             console.log(chalk.red(errStatus));
           }
 
           // это файл?
           if (status.isFile()) {
-            console.log(chalk.white(`\n${inject}.scss - действительно является файлом, добавляем импорт в главный файл стилей...`));
+            console.log(chalk.white(`\n${value}.scss - действительно является файлом, добавляем импорт в главный файл стилей...`));
 
             // Добавим в конец main.scss, импорт файла стилей нашего найденного компонента
             fs.appendFile(injectTo, styleImport, 'utf8', (err) => {
